+++ Assets/com.gamedeveducation.aibehaviours/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.2] - 2 December 2023
+ - Provide a cleaner interface for adding states to the state machine
+ - Provide a cleaner interface for adding state status transition and default transitions
+
 ## [1.0.1] - 23 November 2023
  - Remove incorrect using statement in SMState_Store